src/requirements.txt
flask
flask-redis
Flask-SQLAlchemy
ddtrace==2.8.4
